requirements.txt
boto3
confluent_kafka
pandas
psycopg2-binary
pytest
pylint
python-dotenv
ipykernel